Bug 11431: (QA followup) Fix alter precedences on delete

+++ admin/audio_alerts.pl
@@ -49,7 +49,10 @@ if ( $selector && $sound ) {
     Koha::AudioAlert->new( { selector => $selector, sound => $sound } )->store();
 }
 
-map { Koha::AudioAlerts->find($_)->delete() } @delete;
+if (@delete) {
+    map { Koha::AudioAlerts->find($_)->delete() } @delete;
+    Koha::AudioAlerts->fix_precedences();
+}
 
 if ( $id && $action && $where && $action eq 'move' ) {
     Koha::AudioAlerts->find($id)->move($where);
